Update commit buoi 4 + Bonus Qlkh, Thong ke

+++ src/ra/coursemanagement/business/impl/EnrollmentServiceImpl.java
@@ -3,7 +3,10 @@ package ra.coursemanagement.business.impl;
 import ra.coursemanagement.business.IEnrollmentService;
 import ra.coursemanagement.dao.IEnrollmentDAO;
 import ra.coursemanagement.dao.impl.EnrollmentDAOImpl;
+import ra.coursemanagement.exception.MyCheckedException;
+import ra.coursemanagement.exception.MyUncheckedException;
 import ra.coursemanagement.model.Enrollment;
+import ra.coursemanagement.model.EnrollmentStatus;
 
 import java.util.List;
 
@@ -13,37 +16,107 @@ public class EnrollmentServiceImpl implements IEnrollmentService {
 
     @Override
     public List<Enrollment> findAll() {
-        return enrollmentDAO.findAll();
+
+        try {
+            return enrollmentDAO.findAll();
+        } catch (MyCheckedException e) {
+            throw new MyUncheckedException("Không thể lấy danh sách enrollment: " + e.getMessage());
+        }
     }
 
     @Override
     public Enrollment findById(Integer id) {
-        return enrollmentDAO.findById(id);
+
+        try {
+            return enrollmentDAO.findById(id);
+        } catch (MyCheckedException e) {
+            throw new MyUncheckedException("Không tìm thấy enrollment: " + e.getMessage());
+        }
     }
 
     @Override
     public boolean update(Enrollment e) {
-        return enrollmentDAO.update(e);
+
+        try {
+            Enrollment old = enrollmentDAO.findById(e.getId());
+            if (old == null) {
+                throw new MyUncheckedException("Enrollment không tồn tại!");
+            }
+            if (old.getStatus() == EnrollmentStatus.CONFIRM) {
+                throw new MyUncheckedException("Không thể sửa khóa đã CONFIRM!");
+            }
+            return enrollmentDAO.update(e);
+
+        } catch (MyCheckedException ex) {
+            throw new MyUncheckedException("Cập nhật enrollment thất bại: " + ex.getMessage());
+        }
     }
 
     @Override
     public boolean delete(Integer id) {
-        return enrollmentDAO.delete(id);
-    }
+        try {
+            Enrollment e = enrollmentDAO.findById(id);
+            if (e == null) {
+                throw new MyUncheckedException("Enrollment không tồn tại!");
+            }
+            if (e.getStatus() == EnrollmentStatus.CONFIRM) {
+                throw new MyUncheckedException("Không thể hủy khóa đã CONFIRM!");
+            }
 
+            return enrollmentDAO.delete(id);
+
+        } catch (MyCheckedException ex) {
+            throw new MyUncheckedException("Xóa enrollment thất bại: " + ex.getMessage());
+        }
+    }
 
     @Override
     public List<Enrollment> findByStudentId(int studentId) {
-        return enrollmentDAO.findByStudentId(studentId);
+        try {
+            return enrollmentDAO.findByStudentId(studentId);
+        } catch (MyCheckedException e) {
+            throw new MyUncheckedException("Không thể lấy khóa học của student: " + e.getMessage());
+        }
     }
 
     @Override
     public boolean existsEnrollment(int studentId, int courseId) {
-        return enrollmentDAO.existsEnrollment(studentId,courseId);
+
+        try {
+            return enrollmentDAO.existsEnrollment(studentId, courseId);
+        } catch (MyCheckedException e) {
+            throw new MyUncheckedException("Lỗi kiểm tra enrollment: " + e.getMessage());
+        }
     }
 
     @Override
     public void register(Enrollment e) {
-        enrollmentDAO.save(e);
+
+        try {
+            if (enrollmentDAO.existsEnrollment(e.getStudentId(), e.getCourseId())) {
+                throw new MyUncheckedException("Bạn đã đăng ký khóa học này!");
+            }
+            enrollmentDAO.save(e);
+        } catch (MyCheckedException ex) {
+            throw new MyUncheckedException("Đăng ký khóa học thất bại: " + ex.getMessage());
+        }
+    }
+    @Override
+    public List<Enrollment> findAllAndPaging(int currentPage, int pageSize) {
+
+        try {
+            return enrollmentDAO.findAllAndPaging(currentPage, pageSize);
+        } catch (MyCheckedException e) {
+            throw new MyUncheckedException("Paging enrollment lỗi");
+        }
+    }
+    @Override
+    public List<Enrollment> findByCourseId(int courseId) {
+
+        try {
+            return enrollmentDAO.findByCourseId(courseId);
+        } catch (MyCheckedException e) {
+            throw new MyUncheckedException("Không lấy được enrollment");
+        }
     }
 }
